requirements.txt
# Core dependencies
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
python-multipart>=0.0.6  # For file uploads

# ML dependencies
torch>=2.0.0
torchvision>=0.15.0
Pillow>=10.0.0

# Configuration management  
pydantic>=1.10.0,<2.0.0  # Use v1 for BaseSettings compatibility
python-dotenv>=1.0.0     # For .env file loading

# W&B integration
wandb>=0.16.0

# Optional: Additional utilities
python-json-logger>=2.0.0  # For structured logging
prometheus-client>=0.19.0   # For metrics (optional)

# Development dependencies (optional)
pytest>=7.4.0
pytest-asyncio>=0.21.0
black>=23.0.0
isort>=5.12.0
flake8>=6.0.0
mypy>=1.6.0